Restructed example
